Helicopter Extraction Functionality
Helicopter will now land, open the door, and let player walk up to it, triggering the end of the game. For now it just opens the menu scene. Later the player object will be destroyed and the camera will be moved into the helicopter for a flyover as credits roll.